+++ hass-addon-sunsynk-multi/options.py
@@ -20,18 +20,18 @@ class InverterOptions:
     modbus_id: int = 0
     ha_prefix: str = ""
     serial_nr: str = ""
-    mqtt_id: str = ""
 
     @classmethod
     def factory(cls, opt: dict) -> InverterOptions:
         """Create a class from the options."""
-        serial, _, mid = opt.pop("SERIAL_NR", "").partition(":")
         modbus_id = int(opt.pop("MODBUS_ID"))
-        kwargs = {k.lower(): v for k, v in opt.items()}
-        kwargs["serial_nr"] = serial
-        kwargs["mqtt_id"] = mid or serial
-        kwargs["modbus_id"] = modbus_id
-        return InverterOptions(**kwargs)
+        iopt = InverterOptions(**{k.lower(): v for k, v in opt.items()})
+        iopt.modbus_id = modbus_id
+        if "_" in iopt.serial_nr:
+            _LOGGER.warning(
+                "The serial number contains an underscore '_' - used for testing"
+            )
+        return iopt
 
 
 @attrs.define(slots=True)